fix: ajustes de conflito de tipos de variaveis (decimal e double)

+++ TADS_TP/DTOs/AluguelRequestDTO.cs
@@ -5,7 +5,7 @@
         public DateTime DataInicio { get; set; }
         public DateTime DataFim { get; set; }
         public double QuilometragemInicial { get; set; }
-        public decimal ValorDiaria { get; set; }
+        public double ValorDiaria { get; set; }
         public int VeiculoId { get; set; }
         public int ClienteId { get; set; }
     }
